requirements.txt
google-genai==1.54.0
google-generativeai==0.8.5
openai==2.9.0
anthropic>=0.40.0
jinja2==3.1.6
python-docx==1.2.0
python-dotenv==1.2.1
rich==14.2.0
colorama==0.4.6
chromadb==1.3.7
tiktoken==0.12.0
tenacity==9.1.2
backoff==2.2.1
numpy==2.3.5
pydantic==2.12.5
rank-bm25==0.2.2
sentence-transformers==5.2.0
yt-dlp==2025.12.8
mlx==0.30.0
requests==2.32.5
setuptools>=70.0.0
google-cloud-aiplatform
